Make cuttlefish return the right -config line to node_package
Currently, node_package either generates CONFIG_ARGS line itself or
expects cuttlefish to generate one. In the case of cuttlefish
encountering an existing app.config it simply returned the path to the
app.config file, which is not what node_package expects. This patch
adjusts cuttlefish to return the -config and -args_file string that
would be returned were cuttlefish to be disabled.