Refactor cuttlefish_generator:get_possible_values_for_fuzzy_matches/2.
In the case where `include_default` is used in the schema, and the
user has no entries that match that mapping, the default will be
generated. This presents a problem where, for example, in Riak one
might not be able to disable the HTTP listener because the schema
includes a default one. This commit changes the behavior such that the
`include_default` only affects .conf file generation, not
runtime/start-time app.config generation.