Use ordsets in apply_mappings() fold
The fold function in apply_mappings() would simply cons mappings
to lists for either dropping or keeping a mapping, and then
subtracting the two. The effect was, in a sense, a majority vote:
if more mappings than not indicated dropping, the translation
would be dropped, counter-intuitively.

Solution: make the accumulators ordsets instead. Two test cases
were added: one with 1 'good mapping' and 1 'bad', and the other
with 1 'good' and 2 'bad'. The second test failed before the fix.