Handle case where advanced.config contains > 1 terms
Erlang config files are designed to contain one term, a list, that
contains two-tuples of {Application, Env}. An advanced.config can
possibly contain multiple terms and still parse via file:consult/1, but
cuttlefish does not handle this case. Instead of crashing with
case_clause, we now inform the user that the file is erroneous.