Add cuttlefish_generator:minimal_map/2.
This function generates an Erlang configuration based on the passed
schema that only includes effects from the passed configuration. That
is, defaults in the schema will not be injected and translations for
unspecified settings will not be run.

This allows the caller to coerce, validate and translate incoming
settings for updates to partial configurations, perhaps as part of a
running node.