Repurpose -a to mean --advanced_conf_file as --app_config is not actually used
Currently the -a flag is not used. Originally introduced in [1],
it [seemingly] was meant to allow for app.config location customization.
That never materialized as the flag is not used anywhere in the codebase
besides the getopt definition.

Instead we repurpose -a to mean --advanced_conf_file, which was adopted
in a fork currently bundled with RabbitMQ. Having a hardcoded location
for advanced.config is primarily inconvenient in integration tests
that generate configuration files.

Note that --app_config has been removed. The version will be bumped to
2.3.0.

Per discussion with @dcorbacho @kjnilsson @lukebakken.

1. 
https://github.com/basho/cuttlefish/commit/106b8b769f1ea1f2e9ff3580ab5324e1d8b8d09a